chore: Update project configuration and dependencies
This commit includes updates to the project configuration and dependencies. It adds the `.env.example` file with access key and secret access key for S3, as well as the endpoint and bucket configuration. The `app.module.ts` file is updated to import the `ConfigModule` from `@nestjs/config` and make it global. The `qrcode.module.ts` file is modified to change the order of imports for better readability. The `serverless.yaml` file is updated to change the organization and application names, as well as the bucket name. The `qrcode.service.ts` file is modified to add console logs for debugging purposes. The `package.json` file is updated to add the `@nestjs/config` and `date-fns` dependencies. The `qrcode.controller.ts` file is modified to add error handling and generate a unique key for each uploaded QR code. The `s3.service.ts` file is updated to use the `ConfigService` from `@nestjs/config` for accessing the S3 configuration. The `.serverless/meta.json` file is deleted as it is no longer needed.

```